fix(desktop): split M1 migration tests + document TS ordering domain
Split the M1 harness-migration tests from store_tests.rs into a new
store_migration_tests.rs so store_tests.rs falls back under the 1000-
line limit without a ratchet override; store_tests.rs is now ~858 lines
(checker: 858). The store_tests.rs override is removed; store.rs gains
one line for the new #[path] include (checker: 1044).

Add a one-line comment at sortModelsByKnownTotal noting the identifier
ordering domain is ASCII in practice so UTF-16 vs UTF-8 scalar
divergence for astral code points is accepted.
