+++ desktop/scripts/check-file-sizes.mjs
@@ -84,13 +84,9 @@ const overrides = new Map([
   ["src-tauri/src/archive/mod_tests.rs", 1424],
   // agent-usage-harness: M1 migration (atomic transaction wrapper, PRAGMA
   // guard, canonical-store worklist) + rebuild_metric_index_in_tx helper
-  // + apply_schema_migrations delegating to migrate_add_harness_to_metric_index.
-  // Load-bearing correctness fix (Thufir IMPORTANT). Queued to split.
-  ["src-tauri/src/archive/store.rs", 1043],
-  // agent-usage-harness: three M1 reopen tests (legacy-file open, idempotent
-  // second open, crash-before-commit re-run) via open_archive_db + build_old_schema_db
-  // helper. Test-only content; ratcheted from ~856 -> 1092.
-  ["src-tauri/src/archive/store_tests.rs", 1094],
+  // + apply_schema_migrations + store_migration_tests.rs #[path] include.
+  // Load-bearing correctness fix (Thufir IMPORTANT). M1 tests split out.
+  ["src-tauri/src/archive/store.rs", 1044],
   // unified-agent-model 1A.1: profile reconcile split to agents_profile.rs,
   // ratcheting 1443 -> 1295. Queued to split further in the A2 fold.
   // global-agent-config: resolve_deploy_model_provider + visibility exports
